requirements.txt
aiobotocore==2.3.2
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
alembic==1.8.0
anyio==3.6.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.11.5
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.3.0
bleach==5.0.0
blis==0.7.7
botocore==1.24.21
cachetools==5.2.0
catalogue==2.0.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.1.0
commonmark==0.9.1
coverage==6.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
docker==5.0.3
docutils==0.18.1
e==1.4.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
entrypoints==0.4
eventregistry==8.12
executing==0.8.3
fastjsonschema==2.15.3
-e git+https://github.com/u-amine/femicide_media_watch.git@49cd5a9259c8b5b35b4b503dce70eb3ff393c4be#egg=femicide_media_watch
flake8==2.3.0
Flask==2.1.2
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
funcy==1.17
future==0.18.2
gcsfs==2022.5.0
gensim==4.2.0
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.8.1
google-auth==2.6.6
google-auth-oauthlib==0.5.1
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
iniconfig==1.1.1
ipdb==0.13.9
ipykernel==6.13.0
ipython
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab==3.4.2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.7.0
mistune==0.8.4
nbclassic==0.3.7
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.11
notebook-shim==0.1.0
packaging==21.3
pandas==1.2.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
Pygments==2.12.0
pylint==2.13.9
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.0.0
regex==2022.4.24
requests==2.27.1
scikit-learn
scipy
seaborn
Send2Trash
six
sniffio
soupsieve
stack-data
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
wrapt==1.14.1
zipp==3.8.0

itsdangerous==2.1.2
jmespath==1.0.0
keyring==23.5.1
langcodes==3.3.0
Mako==1.2.0
memoized-property==1.0.3
mlflow==1.26.1
multidict==6.0.2
murmurhash==1.0.7
mypy-extensions==0.4.3
numexpr==2.8.1
numpy==1.20.1
oauthlib==3.2.0
pathspec==0.9.0
pathy==0.6.1
pep8==1.7.1
pkginfo==1.8.2
preshed==3.0.6
prometheus-flask-exporter==0.20.1
protobuf==3.20.1

pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyflakes==2.4.0
PyJWT==2.4.0
pyLDAvis==3.3.1

python-Levenshtein==0.12.2
PyYAML==6.0
querystring-parser==1.2.4
readme-renderer==35.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.4
rsa==4.8
s3fs

sklearn
smart-open
smmap
spacy
spacy-legacy
spacy-loggers
SQLAlchemy
sqlparse
srsly
thinc
threadpoolctl
tinycss2
toml
tomli
tornado
traitlets
twine
typer
typing_extensions
urllib3
wasabi
wcwidth
webencodings
websocket-client
Werkzeug
wrapt
yapf
yarl
zipp
tabulate
termcolor
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs

# utilities
six>=1.14
joblib
memoized-property
termcolor

